Update README.md
Cosnova GmbH Data Science – Use Case:
This use case is meant primarily to demonstrate your skill set and your ability to implement Data Science best practices and to communicate your findings clearly. The factual accuracy is not going to be judged. Please perform a topline analysis of the effect of the COVID-19 pandemic on the beauty industry. The analysis should also contain a predictive component. Also include an executive summary of your findings. Please document your analysis in a Jupyter notebook hosted on GitHub. The Jupyter notebook should be self-contained, i.e., all results should be reproducible by executing the Jupyter notebook or following easy instructions. You may use any publicly available data sources; please include any supplemental material in the GitHub repository. The data collection and implementation within Jupyter should not take longer than 4h. You may make reasonable assumptions; please shortly justify your assumptions.

Dataset and Assumptions
data
The data collected from different data resources of cosmetic e-commerce website.

Assumptions
For the use case: analysis of the effect of the COVID-19 pandemic on the beauty industry.

To analyse Covid effect need the data of sales of the cosmetic products in the year 2020 (Covid effect year sales).

To compare the sales of pandemic year need before year sales data so considered 2019 data also.

This dataset is purely for this perticular usecase and deosn't implies any real data.

In this use case tried to show how the sales of products effected in 2020 compared to 2019.

The prediction model is built to predict the price of the product. Model accurecy came as 93% with Random Forest and 90% with XGBOOST.